+++ MonteCarloMarginalizeCode/Code/RIFT/misc/dag_utils.py
@@ -1944,7 +1944,7 @@ def write_subdagILE_sub(tag='subdag_ile', full_path_name=True, exe=None, univers
     return ile_job, ile_sub_name
 
 
-def write_calibration_uncertainty_reweighting_sub(tag='Calib_reweight', exe=None, log_dir=None, ncopies=1,request_memory=8192,time_marg=True,pickle_file=None,posterior_file=None,universe='vanilla',no_grid=False,ile_args=None,**kwargs):
+def write_calibration_uncertainty_reweighting_sub(tag='Calib_reweight', exe=None, log_dir=None, ncopies=1,request_memory=8192,time_marg=True,pickle_file=None,posterior_file=None,universe='vanilla',no_grid=False,ile_args=None,n_cal=100,**kwargs):
     """
     Write a submit file for launching jobs to reweight final posterior samples due to calibration uncertainty 
 
@@ -1981,7 +1981,7 @@ def write_calibration_uncertainty_reweighting_sub(tag='Calib_reweight', exe=None
     # Add mandatory options
     ile_job.add_opt('data_dump_file', str(pickle_file))
     ile_job.add_opt('posterior_sample_file', str(posterior_file))
-    ile_job.add_opt('number_of_calibration_curves', '100')
+    ile_job.add_opt('number_of_calibration_curves', str(n_cal))
     ile_job.add_opt('reevaluate_likelihood', 'True')
     ile_job.add_opt('use_rift_samples', 'True')
     if time_marg: