+++ MonteCarloMarginalizeCode/Code/bin/combine_weights_and_rejection_sample.py
@@ -23,6 +23,12 @@ def sortKeyFunc(s):
 # load in the arguments
 sample_file = sys.argv[1]
 weights_file_directory = sys.argv[2]
+allow_alternate=False
+allow_duplicates = False
+if len(sys.argv) > 3:   # if 3 arguments passed, target the effective sample size as output size
+    allow_alternate=True
+if len(sys.argv) > 4:
+    allow_duplicates=True
 
 outdir = os.path.dirname(os.path.abspath(sample_file))
 
@@ -55,5 +61,9 @@ if npts_samples > len(weights):
     result.posterior = result.posterior.truncate(after=npts_wts-1)
 
 print(f'Importance sampling efficiency: {np.sum(weights)**2/np.sum(weights**2)/len(weights)}')
-result.posterior = bilby.core.result.rejection_sample(result.posterior, weights)
+if not(allow_alternate):
+    result.posterior = bilby.core.result.rejection_sample(result.posterior, weights)
+else:
+    n_est = int(len(weights)*np.sum(weights)**2/np.sum(weights**2)/len(weights))+1  # effective sample size used to draw
+    result.posterior = result.posterior.sample(n=n_est,weights=weights/np.sum(weights),replace=allow_duplicates)
 result.save_posterior_samples(filename=outdir+'/reweighted_posterior_samples.dat', outdir=outdir)